--- src/main/java/com/jiro4989/tkfm/imageViewer/ImageViewerBorderPaneController.java
@@ -1,240 +1,0 @@
-package com.jiro4989.tkfm.imageViewer;
-
-import com.jiro4989.tkfm.MainController;
-import com.jiro4989.tkfm.model.CroppingImageModel;
-import java.util.Optional;
-import javafx.beans.binding.Bindings;
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.scene.control.ComboBox;
-import javafx.scene.control.Label;
-import javafx.scene.control.Slider;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.input.MouseEvent;
-import javafx.scene.input.ScrollEvent;
-import javafx.scene.layout.GridPane;
-
-public class ImageViewerBorderPaneController {
-  private MainController mainController;
-  private Optional<Image> imageOpt = Optional.empty();
-
-  @FXML private ImageView imageView;
-  @FXML private GridPane imageViewGridPane;
-  @FXML private GridPane focusGridPane;
-
-  @FXML private GridPane trimmingGridPane;
-  @FXML private ImageView trimmingImageView;
-
-  @FXML private Slider slider;
-
-  private ObservableList<Integer> axisItems = FXCollections.observableArrayList(1, 5, 10, 25, 50);
-  @FXML private ComboBox<Integer> axisComboBox;
-  private ObservableList<Integer> zoomRateItems =
-      FXCollections.observableArrayList(1, 5, 10, 25, 50);
-  @FXML private ComboBox<Integer> zoomRateComboBox;
-  @FXML private Label xLabel;
-  @FXML private Label yLabel;
-  @FXML private Label rateLabel;
-
-  private CroppingImageModel cropImage = new CroppingImageModel();
-
-  @FXML
-  private void initialize() {
-    focusGridPane.setOnMouseDragged(e -> moveFocusGridPane(e));
-
-    var pos = cropImage.getPosition();
-    var rect = cropImage.getRectangle();
-
-    Bindings.bindBidirectional(imageView.imageProperty(), cropImage.imageProperty());
-    Bindings.bindBidirectional(focusGridPane.layoutXProperty(), pos.xProperty());
-    Bindings.bindBidirectional(focusGridPane.layoutYProperty(), pos.yProperty());
-    Bindings.bindBidirectional(focusGridPane.prefWidthProperty(), rect.widthProperty());
-    Bindings.bindBidirectional(focusGridPane.prefHeightProperty(), rect.heightProperty());
-    Bindings.bindBidirectional(slider.valueProperty(), cropImage.scaleProperty());
-
-    slider.setOnScroll(e -> changeZoomRateWithScroll(e));
-    slider.valueProperty().addListener(e -> updateImage());
-
-    axisComboBox.setItems(axisItems);
-    zoomRateComboBox.setItems(zoomRateItems);
-    axisComboBox.getSelectionModel().select(1);
-    zoomRateComboBox.getSelectionModel().select(1);
-  }
-
-  public void setMainController(MainController aMainController) {
-    mainController = aMainController;
-  }
-
-  /**
-   * フォーカスをマウスドラッグで動かす。
-   *
-   * @param event マウスイベント
-   */
-  private void moveFocusGridPane(MouseEvent event) {
-    double x = event.getX();
-    double y = event.getY();
-    var rect = cropImage.getRectangle();
-    var w = rect.getWidth();
-    var h = rect.getHeight();
-    var pos = cropImage.getPosition();
-    pos.setX(x - w / 2);
-    pos.setY(y - h / 2);
-    trimmingImageView.setImage(cropImage.crop());
-  }
-
-  /**
-   * フォーカスの移動可能域に制限をかけながら位置移動処理を行う。
-   *
-   * @param x
-   * @param y
-   * @param rate
-   * @param imageWidth
-   * @param imageHeight
-   * @param width
-   */
-  private void setFocusPosition(
-      double x, double y, double rate, double imageWidth, double imageHeight, int width) {
-    x = Math.max(0, x);
-    y = Math.max(0, y);
-    x = Math.min(imageWidth * rate - width, x);
-    y = Math.min(imageHeight * rate - width, y);
-    x = Math.floor(x);
-    y = Math.floor(y);
-
-    focusGridPane.setLayoutX(x);
-    focusGridPane.setLayoutY(y);
-    xLabel.setText("" + x);
-    yLabel.setText("" + y);
-  }
-
-  /**
-   * マウススクロールで拡大率を変更する。
-   *
-   * @param e
-   */
-  private void changeZoomRateWithScroll(ScrollEvent e) {
-    // double deltaY = e.getDeltaY();
-    // double plus = zoomRateComboBox.getSelectionModel().getSelectedItem();
-    // double currentRate = getRate();
-    //
-    // double newRate = 0 < deltaY ? currentRate + plus : currentRate - plus;
-    // newRate = Math.max(0, newRate);
-    // newRate = Math.min(200, newRate);
-    // changeZoomRate(newRate);
-  }
-
-  /** 画像を更新する。 */
-  private void updateImage() {
-    // imageOpt.ifPresent(
-    //     image -> {
-    //       double rate = getRate();
-    //       rate /= 100;
-    //       double imageWidth = image.getWidth();
-    //       double imageHeight = image.getHeight();
-    //       int width = mainController.getTKoolVersion().getWidth();
-    //
-    //       if (width < imageWidth * rate) {
-    //         imageViewGridPane.setPrefSize(imageWidth * rate, imageHeight * rate);
-    //         imageView.setFitWidth(imageWidth * rate);
-    //         imageView.setFitHeight(imageHeight * rate);
-    //
-    //         double x = focusGridPane.getLayoutX();
-    //         double y = focusGridPane.getLayoutY();
-    //         setFocusPosition(x, y, rate, imageWidth, imageHeight, width);
-    //         trimmingImageView.setImage(cropImage.crop());
-    //       }
-    //     });
-  }
-
-  /**
-   * トリミングイメージを返す。
-   *
-   * @return
-   */
-  public Image getTrimmingImage() {
-    return trimmingImageView.getImage();
-  }
-
-  /**
-   * ファイルパスからファイルを取得し、イメージをセットする。
-   *
-   * @param filePath
-   */
-  public void setImage(String filePath) {
-    var img = new Image("file:" + filePath);
-    cropImage.setImage(img);
-    // imageOpt = Optional.ofNullable();
-    // imageOpt.ifPresent(
-    //     image -> {
-    //       double width = image.getWidth();
-    //       double height = image.getHeight();
-    //       imageViewGridPane.setPrefSize(width, height);
-    //       imageView.setImage(image);
-    //       updateImage();
-    //     });
-  }
-
-  public void clearImageView() {
-    imageOpt = Optional.ofNullable(null);
-    imageView.setImage(null);
-    trimmingImageView.setImage(null);
-  }
-
-  public void moveUp() {
-    double n = axisComboBox.getSelectionModel().getSelectedItem();
-    cropImage.moveUp(n);
-  }
-
-  public void moveLeft() {
-    double n = axisComboBox.getSelectionModel().getSelectedItem();
-    cropImage.moveLeft(n);
-  }
-
-  public void moveDown() {
-    double n = axisComboBox.getSelectionModel().getSelectedItem();
-    cropImage.moveDown(n);
-  }
-
-  public void moveRight() {
-    double n = axisComboBox.getSelectionModel().getSelectedItem();
-    cropImage.moveRight(n);
-  }
-
-  public double getX() {
-    return Double.valueOf(xLabel.getText());
-  }
-
-  public double getY() {
-    return Double.valueOf(yLabel.getText());
-  }
-
-  public double getRate() {
-    return Double.valueOf(rateLabel.getText());
-  }
-
-  public void changeVersion(double width) {
-    // focusGridPane.setPrefSize(width, width);
-    // trimmingGridPane.setPrefSize(width, width);
-    // trimmingImageView.setFitWidth(width);
-    // trimmingImageView.setFitHeight(width);
-    // trimmingImageView.setImage(cropImage.crop());
-  }
-
-  public void zoomIn() {
-    // double rate = getRate();
-    // double plus = zoomRateComboBox.getSelectionModel().getSelectedItem();
-    // changeZoomRate(rate + plus);
-  }
-
-  public void zoomOut() {
-    // double rate = getRate();
-    // double plus = zoomRateComboBox.getSelectionModel().getSelectedItem();
-    // changeZoomRate(rate - plus);
-  }
-
-  public CroppingImageModel getCroppingImageModel() {
-    return cropImage;
-  }
-}
