--- src/main/resources/com/jiro4989/tkfm/imageViewer/ImageViewerBorderPane.fxml
@@ -1,113 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.lang.*?>
-<?import javafx.collections.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.image.*?>
-<?import javafx.scene.layout.*?>
-<?import javafx.scene.layout.BorderPane?>
-
-<BorderPane prefHeight="478.0" prefWidth="862.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/2.2" fx:controller="com.jiro4989.tkfm.imageViewer.ImageViewerBorderPaneController">
-  <!-- TODO Add Nodes -->
-  <bottom>
-    <ToolBar BorderPane.alignment="CENTER">
-      <items>
-        <Label text="Ｘ：" />
-        <Label fx:id="xLabel" prefWidth="75.0" text="0.0" />
-        <Label text="Ｙ：" />
-        <Label fx:id="yLabel" prefWidth="75.0" text="0.0" />
-        <Separator orientation="VERTICAL" prefHeight="14.0" />
-        <Label text="移動幅：" />
-        <ComboBox fx:id="axisComboBox" prefWidth="100.0">
-          <items>
-            <FXCollections fx:factory="observableArrayList">
-              <String fx:value="アイテム1" />
-              <String fx:value="アイテム2" />
-              <String fx:value="アイテム3" />
-            </FXCollections>
-          </items>
-        </ComboBox>
-        <Separator orientation="VERTICAL" prefHeight="14.0" />
-        <Label text="拡大率：" />
-        <Label fx:id="rateLabel" prefWidth="75.0" text="100.0" />
-        <Label text="拡大幅：" />
-        <ComboBox fx:id="zoomRateComboBox" prefWidth="100.0">
-          <items>
-            <FXCollections fx:factory="observableArrayList">
-              <String fx:value="アイテム1" />
-              <String fx:value="アイテム2" />
-              <String fx:value="アイテム3" />
-            </FXCollections>
-          </items>
-        </ComboBox>
-      </items>
-      <BorderPane.margin>
-        <Insets />
-      </BorderPane.margin>
-    </ToolBar>
-  </bottom>
-  <center>
-    <BorderPane prefHeight="200.0" prefWidth="200.0">
-      <bottom>
-        <Slider fx:id="slider" blockIncrement="5.0" majorTickUnit="25.0" max="200.0" minorTickCount="4" showTickLabels="true" showTickMarks="true" snapToTicks="true" value="100.0" />
-      </bottom>
-      <center>
-        <ScrollPane id="imageRootScroolPane" prefHeight="425.0" prefWidth="659.0">
-          <content>
-            <AnchorPane id="AnchorPane">
-              <children>
-                <GridPane id="GridPane" fx:id="imageViewGridPane" gridLinesVisible="true" AnchorPane.leftAnchor="0.0" AnchorPane.topAnchor="0.0">
-                  <children>
-                    <ImageView fx:id="imageView" fitHeight="417.75" fitWidth="557.0" pickOnBounds="true" preserveRatio="true" GridPane.columnIndex="0" GridPane.rowIndex="0" />
-                  </children>
-                  <columnConstraints>
-                    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" />
-                  </columnConstraints>
-                  <rowConstraints>
-                    <RowConstraints minHeight="10.0" vgrow="SOMETIMES" />
-                  </rowConstraints>
-                </GridPane>
-                <GridPane fx:id="focusGridPane" gridLinesVisible="true" layoutX="207.0" layoutY="160.0" prefHeight="144.0" prefWidth="144.0">
-                  <columnConstraints>
-                    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                  </columnConstraints>
-                  <rowConstraints>
-                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                  </rowConstraints>
-                </GridPane>
-              </children>
-            </AnchorPane>
-          </content>
-          <BorderPane.margin>
-            <Insets bottom="5.0" left="5.0" right="5.0" top="5.0" />
-          </BorderPane.margin>
-        </ScrollPane>
-      </center>
-      <left>
-        <AnchorPane prefHeight="435.0" prefWidth="163.0">
-          <children>
-            <FlowPane alignment="BOTTOM_LEFT" columnHalignment="LEFT" orientation="HORIZONTAL" prefHeight="200.0" prefWidth="149.0" rowValignment="BOTTOM" AnchorPane.bottomAnchor="5.0" AnchorPane.leftAnchor="10.0">
-              <children>
-                <Label id="" prefHeight="30.0" prefWidth="106.0" text="トリミング：" />
-                <GridPane fx:id="trimmingGridPane" gridLinesVisible="true" prefHeight="144.0" prefWidth="144.0">
-                  <children>
-                    <ImageView fx:id="trimmingImageView" fitHeight="144.0" fitWidth="144.0" pickOnBounds="true" preserveRatio="true" GridPane.columnIndex="0" GridPane.rowIndex="0" />
-                  </children>
-                  <columnConstraints>
-                    <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                  </columnConstraints>
-                  <rowConstraints>
-                    <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                  </rowConstraints>
-                </GridPane>
-              </children>
-            </FlowPane>
-          </children>
-        </AnchorPane>
-      </left>
-    </BorderPane>
-  </center>
-</BorderPane>
